pkg tests, reconnect tests moved
