+++ natspkg_test.go
@@ -1,0 +1,230 @@
+package nats
+
+////////////////////////////////////////////////////////////////////////////////
+// Package scoped specific tests here..
+////////////////////////////////////////////////////////////////////////////////
+
+import (
+	"errors"
+	"reflect"
+	"testing"
+	"time"
+
+	"github.com/nats-io/gnatsd/server"
+	gnatsd "github.com/nats-io/gnatsd/test"
+)
+
+// Dumb wait program to sync on callbacks, etc... Will timeout
+func Wait(ch chan bool) error {
+	return WaitTime(ch, 500*time.Millisecond)
+}
+
+func WaitTime(ch chan bool, timeout time.Duration) error {
+	select {
+	case <-ch:
+		return nil
+	case <-time.After(timeout):
+	}
+	return errors.New("timeout")
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Reconnect tests
+////////////////////////////////////////////////////////////////////////////////
+
+var reconnectOpts = Options{
+	Url:            "nats://localhost:22222",
+	AllowReconnect: true,
+	MaxReconnect:   10,
+	ReconnectWait:  100 * time.Millisecond,
+	Timeout:        DefaultTimeout,
+}
+
+func RunServerOnPort(port int) *server.Server {
+	opts := gnatsd.DefaultTestOptions
+	opts.Port = port
+	return RunServerWithOptions(opts)
+}
+
+func RunServerWithOptions(opts server.Options) *server.Server {
+	return gnatsd.RunServer(&opts)
+}
+
+func TestReconnectServerStats(t *testing.T) {
+	ts := RunServerOnPort(22222)
+
+	opts := reconnectOpts
+	nc, _ := opts.Connect()
+	nc.Flush()
+
+	ts.Shutdown()
+	// server is stopped here...
+
+	ts = RunServerOnPort(22222)
+	defer ts.Shutdown()
+
+	if err := nc.FlushTimeout(5 * time.Second); err != nil {
+		t.Fatalf("Error on Flush: %v", err)
+	}
+
+	// Make sure the server who is reconnected has the reconnects stats reset.
+	_, cur := nc.currentServer()
+	if cur.reconnects != 0 {
+		t.Fatalf("Current Server's reconnects should be 0 vs %d\n", cur.reconnects)
+	}
+	nc.Close()
+}
+
+func TestParseStateReconnectFunctionality(t *testing.T) {
+	ts := RunServerOnPort(22222)
+	ch := make(chan bool)
+
+	opts := reconnectOpts
+	nc, _ := opts.Connect()
+	ec, err := NewEncodedConn(nc, DEFAULT_ENCODER)
+	if err != nil {
+		t.Fatalf("Failed to create an encoded connection: %v\n", err)
+	}
+
+	testString := "bar"
+	ec.Subscribe("foo", func(s string) {
+		if s != testString {
+			t.Fatal("String doesn't match")
+		}
+		ch <- true
+	})
+	ec.Flush()
+
+	// Simulate partialState, this needs to be cleared
+	nc.mu.Lock()
+	nc.ps.state = OP_PON
+	nc.mu.Unlock()
+
+	ts.Shutdown()
+	// server is stopped here...
+
+	dch := make(chan bool)
+	opts.DisconnectedCB = func(_ *Conn) {
+		dch <- true
+	}
+	Wait(dch)
+
+	if err := ec.Publish("foo", testString); err != nil {
+		t.Fatalf("Failed to publish message: %v\n", err)
+	}
+
+	ts = RunServerOnPort(22222)
+	defer ts.Shutdown()
+
+	if err := ec.FlushTimeout(5 * time.Second); err != nil {
+		t.Fatalf("Error on Flush: %v", err)
+	}
+
+	if e := Wait(ch); e != nil {
+		t.Fatal("Did not receive our message")
+	}
+
+	expectedReconnectCount := uint64(1)
+	if ec.Conn.Reconnects != expectedReconnectCount {
+		t.Fatalf("Reconnect count incorrect: %d vs %d\n",
+			ec.Conn.Reconnects, expectedReconnectCount)
+	}
+
+	nc.Close()
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// ServerPool tests
+////////////////////////////////////////////////////////////////////////////////
+
+var testServers = []string{
+	"nats://localhost:1222",
+	"nats://localhost:1223",
+	"nats://localhost:1224",
+	"nats://localhost:1225",
+	"nats://localhost:1226",
+	"nats://localhost:1227",
+	"nats://localhost:1228",
+}
+
+func TestServersRandomize(t *testing.T) {
+	opts := DefaultOptions
+	opts.Servers = testServers
+	nc := &Conn{Opts: opts}
+	if err := nc.setupServerPool(); err != nil {
+		t.Fatalf("Problem setting up Server Pool: %v\n", err)
+	}
+	// Build []string from srvPool
+	clientServers := []string{}
+	for _, s := range nc.srvPool {
+		clientServers = append(clientServers, s.url.String())
+	}
+	// In theory this could happen..
+	if reflect.DeepEqual(testServers, clientServers) {
+		t.Fatalf("ServerPool list not randomized\n")
+	}
+
+	// Now test that we do not randomize if proper flag is set.
+	opts = DefaultOptions
+	opts.Servers = testServers
+	opts.NoRandomize = true
+	nc = &Conn{Opts: opts}
+	if err := nc.setupServerPool(); err != nil {
+		t.Fatalf("Problem setting up Server Pool: %v\n", err)
+	}
+	// Build []string from srvPool
+	clientServers = []string{}
+	for _, s := range nc.srvPool {
+		clientServers = append(clientServers, s.url.String())
+	}
+	if !reflect.DeepEqual(testServers, clientServers) {
+		t.Fatalf("ServerPool list should not be randomized\n")
+	}
+}
+
+func TestSelectNextServer(t *testing.T) {
+	opts := DefaultOptions
+	opts.Servers = testServers
+	opts.NoRandomize = true
+	nc := &Conn{Opts: opts}
+	if err := nc.setupServerPool(); err != nil {
+		t.Fatalf("Problem setting up Server Pool: %v\n", err)
+	}
+	if nc.url != nc.srvPool[0].url {
+		t.Fatalf("Wrong default selection: %v\n", nc.url)
+	}
+
+	sel, err := nc.selectNextServer()
+	if err != nil {
+		t.Fatalf("Got an err: %v\n", err)
+	}
+	// Check that we are now looking at #2, and current is now last.
+	if len(nc.srvPool) != len(testServers) {
+		t.Fatalf("List is incorrect size: %d vs %d\n", len(nc.srvPool), len(testServers))
+	}
+	if nc.url.String() != testServers[1] {
+		t.Fatalf("Selection incorrect: %v vs %v\n", nc.url, testServers[1])
+	}
+	if nc.srvPool[len(nc.srvPool)-1].url.String() != testServers[0] {
+		t.Fatalf("Did not push old to last position\n")
+	}
+	if sel != nc.srvPool[0] {
+		t.Fatalf("Did not return correct server: %v vs %v\n", sel.url, nc.srvPool[0].url)
+	}
+
+	// Test that we do not keep servers where we have tried to reconnect past our limit.
+	nc.srvPool[0].reconnects = int(opts.MaxReconnect)
+	if _, err := nc.selectNextServer(); err != nil {
+		t.Fatalf("Got an err: %v\n", err)
+	}
+	// Check that we are now looking at #3, and current is not in the list.
+	if len(nc.srvPool) != len(testServers)-1 {
+		t.Fatalf("List is incorrect size: %d vs %d\n", len(nc.srvPool), len(testServers)-1)
+	}
+	if nc.url.String() != testServers[2] {
+		t.Fatalf("Selection incorrect: %v vs %v\n", nc.url, testServers[2])
+	}
+	if nc.srvPool[len(nc.srvPool)-1].url.String() == testServers[1] {
+		t.Fatalf("Did not throw away the last server correctly\n")
+	}
+}
