+++ test/reconnect_test.go
@@ -1,18 +1,21 @@
-package nats
+package test
 
 import (
 	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
+
+	"github.com/nats-io/gnatsd/server"
+	"github.com/nats-io/nats"
 )
 
-func startReconnectServer(t *testing.T) *server {
-	return startServer(t, 22222, "")
+func startReconnectServer(t *testing.T) *server.Server {
+	return RunServerOnPort(22222)
 }
 
 func TestReconnectTotalTime(t *testing.T) {
-	opts := DefaultOptions
+	opts := nats.DefaultOptions
 	totalReconnectTime := time.Duration(opts.MaxReconnect) * opts.ReconnectWait
 	if totalReconnectTime < (2 * time.Minute) {
 		t.Fatalf("Total reconnect time should be at least 2 mins: Currently %v\n",
@@ -23,10 +26,10 @@ func TestReconnectTotalTime(t *testing.T) {
 func TestReconnectDisallowedFlags(t *testing.T) {
 	ts := startReconnectServer(t)
 	ch := make(chan bool)
-	opts := DefaultOptions
+	opts := nats.DefaultOptions
 	opts.Url = "nats://localhost:22222"
 	opts.AllowReconnect = false
-	opts.ClosedCB = func(_ *Conn) {
+	opts.ClosedCB = func(_ *nats.Conn) {
 		ch <- true
 	}
 	nc, err := opts.Connect()
@@ -34,8 +37,9 @@ func TestReconnectDisallowedFlags(t *testing.T) {
 		t.Fatalf("Should have connected ok: %v", err)
 	}
 
-	ts.stopServer()
-	if e := wait(ch); e != nil {
+	ts.Shutdown()
+
+	if e := Wait(ch); e != nil {
 		t.Fatal("Did not trigger ClosedCB correctly")
 	}
 	nc.Close()
@@ -43,28 +47,29 @@ func TestReconnectDisallowedFlags(t *testing.T) {
 
 func TestReconnectAllowedFlags(t *testing.T) {
 	ts := startReconnectServer(t)
+	defer ts.Shutdown()
 	ch := make(chan bool)
-	opts := DefaultOptions
+	opts := nats.DefaultOptions
 	opts.Url = "nats://localhost:22222"
 	opts.AllowReconnect = true
 	opts.MaxReconnect = 2
 	opts.ReconnectWait = 1 * time.Second
 
-	opts.ClosedCB = func(_ *Conn) {
+	opts.ClosedCB = func(_ *nats.Conn) {
 		ch <- true
 	}
 	nc, err := opts.Connect()
 	if err != nil {
 		t.Fatalf("Should have connected ok: %v", err)
 	}
-	ts.stopServer()
+	ts.Shutdown()
 
 	// We want wait to timeout here, and the connection
 	// should not trigger the Close CB.
-	if e := wait(ch); e == nil {
+	if e := Wait(ch); e == nil {
 		t.Fatal("Triggered ClosedCB incorrectly")
 	}
-	if !nc.isReconnecting() {
+	if !nc.IsReconnecting() {
 		t.Fatal("Expected to be in a reconnecting state")
 	}
 
@@ -73,21 +78,22 @@ func TestReconnectAllowedFlags(t *testing.T) {
 	nc.Close()
 }
 
-var reconnectOpts = Options{
+var reconnectOpts = nats.Options{
 	Url:            "nats://localhost:22222",
 	AllowReconnect: true,
 	MaxReconnect:   10,
 	ReconnectWait:  100 * time.Millisecond,
-	Timeout:        DefaultTimeout,
+	Timeout:        nats.DefaultTimeout,
 }
 
 func TestBasicReconnectFunctionality(t *testing.T) {
 	ts := startReconnectServer(t)
+
 	ch := make(chan bool)
 
 	opts := reconnectOpts
 	nc, _ := opts.Connect()
-	ec, err := NewEncodedConn(nc, "default")
+	ec, err := nats.NewEncodedConn(nc, nats.DEFAULT_ENCODER)
 	if err != nil {
 		t.Fatalf("Failed to create an encoded connection: %v\n", err)
 	}
@@ -101,27 +107,27 @@ func TestBasicReconnectFunctionality(t *testing.T) {
 	})
 	ec.Flush()
 
-	ts.stopServer()
+	ts.Shutdown()
 	// server is stopped here...
 
 	dch := make(chan bool)
-	opts.DisconnectedCB = func(_ *Conn) {
+	opts.DisconnectedCB = func(_ *nats.Conn) {
 		dch <- true
 	}
-	wait(dch)
+	Wait(dch)
 
 	if err := ec.Publish("foo", testString); err != nil {
 		t.Fatalf("Failed to publish message: %v\n", err)
 	}
 
 	ts = startReconnectServer(t)
-	defer ts.stopServer()
+	defer ts.Shutdown()
 
 	if err := ec.FlushTimeout(5 * time.Second); err != nil {
 		t.Fatalf("Error on Flush: %v", err)
 	}
 
-	if e := wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		t.Fatal("Did not receive our message")
 	}
 
@@ -130,13 +136,6 @@ func TestBasicReconnectFunctionality(t *testing.T) {
 		t.Fatalf("Reconnect count incorrect: %d vs %d\n",
 			ec.Conn.Reconnects, expectedReconnectCount)
 	}
-
-	// Make sure the server who is reconnected has the reconnects stats reset.
-	_, cur := nc.currentServer()
-	if cur.reconnects != 0 {
-		t.Fatalf("Current Server's reconnects should be 0 vs %d\n", cur.reconnects)
-	}
-
 	nc.Close()
 }
 
@@ -145,18 +144,18 @@ func TestExtendedReconnectFunctionality(t *testing.T) {
 
 	opts := reconnectOpts
 	dch := make(chan bool)
-	opts.DisconnectedCB = func(_ *Conn) {
+	opts.DisconnectedCB = func(_ *nats.Conn) {
 		dch <- true
 	}
 	rch := make(chan bool)
-	opts.ReconnectedCB = func(_ *Conn) {
+	opts.ReconnectedCB = func(_ *nats.Conn) {
 		rch <- true
 	}
 	nc, err := opts.Connect()
 	if err != nil {
 		t.Fatalf("Should have connected ok: %v", err)
 	}
-	ec, err := NewEncodedConn(nc, "default")
+	ec, err := nats.NewEncodedConn(nc, nats.DEFAULT_ENCODER)
 	if err != nil {
 		t.Fatalf("Failed to create an encoded connection: %v\n", err)
 	}
@@ -174,11 +173,11 @@ func TestExtendedReconnectFunctionality(t *testing.T) {
 	ec.Publish("foo", testString)
 	ec.Flush()
 
-	ts.stopServer()
+	ts.Shutdown()
 	// server is stopped here..
 
 	// wait for disconnect
-	if e := waitTime(dch, 2*time.Second); e != nil {
+	if e := WaitTime(dch, 2*time.Second); e != nil {
 		t.Fatal("Did not receive a disconnect callback message")
 	}
 
@@ -199,11 +198,11 @@ func TestExtendedReconnectFunctionality(t *testing.T) {
 	}
 
 	ts = startReconnectServer(t)
-	defer ts.stopServer()
+	defer ts.Shutdown()
 
 	// server is restarted here..
 	// wait for reconnect
-	if e := waitTime(rch, 2*time.Second); e != nil {
+	if e := WaitTime(rch, 2*time.Second); e != nil {
 		t.Fatal("Did not receive a reconnect callback message")
 	}
 
@@ -221,71 +220,13 @@ func TestExtendedReconnectFunctionality(t *testing.T) {
 	})
 	ec.Publish("done", true)
 
-	if e := wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		t.Fatal("Did not receive our message")
 	}
 
 	if atomic.LoadInt32(&received) != 4 {
 		t.Fatalf("Received != %d, equals %d\n", 4, received)
 	}
-}
-
-func TestParseStateReconnectFunctionality(t *testing.T) {
-	ts := startReconnectServer(t)
-	ch := make(chan bool)
-
-	opts := reconnectOpts
-	nc, _ := opts.Connect()
-	ec, err := NewEncodedConn(nc, "default")
-	if err != nil {
-		t.Fatalf("Failed to create an encoded connection: %v\n", err)
-	}
-
-	testString := "bar"
-	ec.Subscribe("foo", func(s string) {
-		if s != testString {
-			t.Fatal("String doesn't match")
-		}
-		ch <- true
-	})
-	ec.Flush()
-
-	// Simulate partialState, this needs to be cleared
-	nc.mu.Lock()
-	nc.ps.state = OP_PON
-	nc.mu.Unlock()
-
-	ts.stopServer()
-	// server is stopped here...
-
-	dch := make(chan bool)
-	opts.DisconnectedCB = func(_ *Conn) {
-		dch <- true
-	}
-	wait(dch)
-
-	if err := ec.Publish("foo", testString); err != nil {
-		t.Fatalf("Failed to publish message: %v\n", err)
-	}
-
-	ts = startReconnectServer(t)
-	defer ts.stopServer()
-
-	if err := ec.FlushTimeout(5 * time.Second); err != nil {
-		t.Fatalf("Error on Flush: %v", err)
-	}
-
-	if e := wait(ch); e != nil {
-		t.Fatal("Did not receive our message")
-	}
-
-	expectedReconnectCount := uint64(1)
-	if ec.Conn.Reconnects != expectedReconnectCount {
-		t.Fatalf("Reconnect count incorrect: %d vs %d\n",
-			ec.Conn.Reconnects, expectedReconnectCount)
-	}
-
-	nc.Close()
 }
 
 func TestQueueSubsOnReconnect(t *testing.T) {
@@ -295,7 +236,7 @@ func TestQueueSubsOnReconnect(t *testing.T) {
 
 	// Allow us to block on reconnect complete.
 	reconnectsDone := make(chan bool)
-	opts.ReconnectedCB = func(nc *Conn) {
+	opts.ReconnectedCB = func(nc *nats.Conn) {
 		reconnectsDone <- true
 	}
 
@@ -311,7 +252,7 @@ func TestQueueSubsOnReconnect(t *testing.T) {
 
 	// Create connection
 	nc, _ := opts.Connect()
-	ec, err := NewEncodedConn(nc, "json")
+	ec, err := nats.NewEncodedConn(nc, nats.JSON_ENCODER)
 	if err != nil {
 		t.Fatalf("Failed to create an encoded connection: %v\n", err)
 	}
@@ -367,9 +308,9 @@ func TestQueueSubsOnReconnect(t *testing.T) {
 	sendAndCheckMsgs(10)
 
 	// Stop and restart server
-	ts.stopServer()
+	ts.Shutdown()
 	ts = startReconnectServer(t)
-	defer ts.stopServer()
+	defer ts.Shutdown()
 
 	waitOnReconnect()
 
@@ -379,11 +320,11 @@ func TestQueueSubsOnReconnect(t *testing.T) {
 
 func TestIsClosed(t *testing.T) {
 	ts := startReconnectServer(t)
-	nc := newConnection(t)
+	nc := NewConnection(t, 22222)
 	if nc.IsClosed() == true {
 		t.Fatalf("IsClosed returned true when the connection is still open.")
 	}
-	ts.stopServer()
+	ts.Shutdown()
 	if nc.IsClosed() == true {
 		t.Fatalf("IsClosed returned true when the connection is still open.")
 	}
@@ -395,25 +336,25 @@ func TestIsClosed(t *testing.T) {
 	if nc.IsClosed() == false {
 		t.Fatalf("IsClosed returned false after Close() was called.")
 	}
-	ts.stopServer()
+	ts.Shutdown()
 }
 
 func TestIsReconnectingAndStatus(t *testing.T) {
 	ts := startReconnectServer(t)
 	// This will kill the last 'ts' server that is created
-	defer func() { ts.stopServer() }()
+	defer func() { ts.Shutdown() }()
 	disconnectedch := make(chan bool)
 	reconnectch := make(chan bool)
-	opts := DefaultOptions
+	opts := nats.DefaultOptions
 	opts.Url = "nats://localhost:22222"
 	opts.AllowReconnect = true
 	opts.MaxReconnect = 10000
 	opts.ReconnectWait = 100 * time.Millisecond
 
-	opts.DisconnectedCB = func(_ *Conn) {
+	opts.DisconnectedCB = func(_ *nats.Conn) {
 		disconnectedch <- true
 	}
-	opts.ReconnectedCB = func(_ *Conn) {
+	opts.ReconnectedCB = func(_ *nats.Conn) {
 		reconnectch <- true
 	}
 
@@ -425,32 +366,32 @@ func TestIsReconnectingAndStatus(t *testing.T) {
 	if nc.IsReconnecting() == true {
 		t.Fatalf("IsReconnecting returned true when the connection is still open.")
 	}
-	if status := nc.Status(); status != CONNECTED {
+	if status := nc.Status(); status != nats.CONNECTED {
 		t.Fatalf("Status returned %d when connected instead of CONNECTED", status)
 	}
-	ts.stopServer()
+	ts.Shutdown()
 
 	// Wait until we get the disconnected callback
-	if e := wait(disconnectedch); e != nil {
+	if e := Wait(disconnectedch); e != nil {
 		t.Fatalf("Disconnect callback wasn't triggered: %v", e)
 	}
 	if nc.IsReconnecting() == false {
 		t.Fatalf("IsReconnecting returned false when the client is reconnecting.")
 	}
-	if status := nc.Status(); status != RECONNECTING {
+	if status := nc.Status(); status != nats.RECONNECTING {
 		t.Fatalf("Status returned %d when reconnecting instead of CONNECTED", status)
 	}
 
 	ts = startReconnectServer(t)
 
 	// Wait until we get the reconnect callback
-	if e := wait(reconnectch); e != nil {
+	if e := Wait(reconnectch); e != nil {
 		t.Fatalf("Reconnect callback wasn't triggered: %v", e)
 	}
 	if nc.IsReconnecting() == true {
 		t.Fatalf("IsReconnecting returned true after the connection was reconnected.")
 	}
-	if status := nc.Status(); status != CONNECTED {
+	if status := nc.Status(); status != nats.CONNECTED {
 		t.Fatalf("Status returned %d when reconnected instead of CONNECTED", status)
 	}
 
@@ -459,7 +400,7 @@ func TestIsReconnectingAndStatus(t *testing.T) {
 	if nc.IsReconnecting() == true {
 		t.Fatalf("IsReconnecting returned true after Close() was called.")
 	}
-	if status := nc.Status(); status != CLOSED {
+	if status := nc.Status(); status != nats.CLOSED {
 		t.Fatalf("Status returned %d after Close() was called instead of CLOSED", status)
 	}
 }
